move around to sort and shim
