+++ packages/client/src/gen/coordinator/index.ts
@@ -1,5 +1,8 @@
 /* tslint:disable */
 /* eslint-disable */
+
+import { Bound, DailyAggregateCallStatsResponse } from "../shims";
+
 /**
  *
  * @export
@@ -491,6 +494,32 @@ export interface CallDeletedEvent {
   type: string;
 }
 /**
+ *
+ * @export
+ * @interface CallDurationReportResponse
+ */
+export interface CallDurationReportResponse {
+  /**
+   *
+   * @type {Array<DailyAggregateCallStatsResponse<CallDurationReport>>}
+   * @memberof CallDurationReportResponse
+   */
+  daily: Array<DailyAggregateCallStatsResponse<CallDurationReport>>;
+}
+/**
+ *
+ * @export
+ * @interface CallDurationReport
+ */
+export interface CallDurationReport {
+  /**
+   *
+   * @type {Array<ReportByHistogramBucket>}
+   * @memberof CallDurationReport
+   */
+  histogram: Array<ReportByHistogramBucket>;
+}
+/**
  * This event is sent when a call is mark as ended for all its participants. Clients receiving this event should leave the call screen
  * @export
  * @interface CallEndedEvent
@@ -977,6 +1006,32 @@ export interface CallParticipantResponse {
    * @memberof CallParticipantResponse
    */
   user_session_id: string;
+}
+/**
+ *
+ * @export
+ * @interface CallParticipantCountReportResponse
+ */
+export interface CallParticipantCountReportResponse {
+  /**
+   *
+   * @type {Array<DailyAggregateCallStatsResponse<CallParticipantCountReport>>}
+   * @memberof CallParticipantCountReportResponse
+   */
+  daily: Array<DailyAggregateCallStatsResponse<CallParticipantCountReport>>;
+}
+/**
+ *
+ * @export
+ * @interface CallParticipantCountReport
+ */
+export interface CallParticipantCountReport {
+  /**
+   *
+   * @type {Array<ReportByHistogramBucket>}
+   * @memberof CallParticipantCountReport
+   */
+  histogram: Array<ReportByHistogramBucket>;
 }
 /**
  * This event is sent when a reaction is sent in a call, clients should use this to show the reaction in the call screen
@@ -1836,6 +1891,32 @@ export interface CallSettingsResponse {
    * @memberof CallSettingsResponse
    */
   video: VideoSettingsResponse;
+}
+/**
+ *
+ * @export
+ * @interface CallsPerDayReportResponse
+ */
+export interface CallsPerDayReportResponse {
+  /**
+   *
+   * @type {Array<DailyAggregateCallStatsResponse<CallsPerDayReport>>}
+   * @memberof CallsPerDayReportResponse
+   */
+  daily: Array<DailyAggregateCallStatsResponse<CallsPerDayReport>>;
+}
+/**
+ *
+ * @export
+ * @interface CallsPerDayReport
+ */
+export interface CallsPerDayReport {
+  /**
+   *
+   * @type {number}
+   * @memberof CallsPerDayReport
+   */
+  count: number;
 }
 /**
  *
@@ -4226,6 +4307,37 @@ export interface MuteUsersResponse {
 /**
  *
  * @export
+ * @interface NetworkMetricsReportResponse
+ */
+export interface NetworkMetricsReportResponse {
+  /**
+   *
+   * @type {number}
+   * @memberof NetworkMetricsReportResponse
+   */
+  average_connection_time?: number;
+  /**
+   *
+   * @type {number}
+   * @memberof NetworkMetricsReportResponse
+   */
+  average_jitter?: number;
+  /**
+   *
+   * @type {number}
+   * @memberof NetworkMetricsReportResponse
+   */
+  average_latency?: number;
+  /**
+   *
+   * @type {number}
+   * @memberof NetworkMetricsReportResponse
+   */
+  average_time_to_reconnect?: number;
+}
+/**
+ *
+ * @export
  * @interface NoiseCancellationSettings
  */
 export interface NoiseCancellationSettings {
@@ -4676,6 +4788,25 @@ export interface PermissionRequestEvent {
 /**
  *
  * @export
+ * @interface PerSDKUsageReport
+ */
+export interface PerSDKUsageReport {
+  /**
+   *
+   * @type {{ [key: string]: number; }}
+   * @memberof PerSDKUsageReport
+   */
+  by_version: { [key: string]: number };
+  /**
+   *
+   * @type {number}
+   * @memberof PerSDKUsageReport
+   */
+  total: number;
+}
+/**
+ *
+ * @export
  * @interface PinRequest
  */
 export interface PinRequest {
@@ -4826,7 +4957,32 @@ export interface PushNotificationSettingsResponse {
    */
   disabled_until?: string;
 }
-
+/**
+ *
+ * @export
+ * @interface QualityScoreReportResponse
+ */
+export interface QualityScoreReportResponse {
+  /**
+   *
+   * @type {Array<DailyAggregateCallStatsResponse<QualityScoreReport>>}
+   * @memberof QualityScoreReportResponse
+   */
+  daily: Array<DailyAggregateCallStatsResponse<QualityScoreReport>>;
+}
+/**
+ *
+ * @export
+ * @interface QualityScoreReport
+ */
+export interface QualityScoreReport {
+  /**
+   *
+   * @type {Array<ReportByHistogramBucket>}
+   * @memberof QualityScoreReport
+   */
+  histogram: Array<ReportByHistogramBucket>;
+}
 /**
  *
  * @export
@@ -4852,316 +5008,6 @@ export interface QueryAggregateCallStatsRequest {
    */
   to?: string;
 }
-
-/**
- *
- * @export
- * @interface CallDurationReportResponse
- */
-export interface CallDurationReportResponse {
-  /**
-   *
-   * @type {Array<DailyAggregateCallStatsResponse<CallDurationReport>>}
-   * @memberof CallDurationReportResponse
-   */
-  daily: Array<DailyAggregateCallStatsResponse<CallDurationReport>>;
-}
-/**
- *
- * @export
- * @interface CallDurationReport
- */
-export interface CallDurationReport {
-  /**
-   *
-   * @type {Array<ReportByHistogramBucket>}
-   * @memberof CallDurationReport
-   */
-  histogram: Array<ReportByHistogramBucket>;
-}
-
-/**
- *
- * @export
- * @interface CallParticipantCountReportResponse
- */
-export interface CallParticipantCountReportResponse {
-  /**
-   *
-   * @type {Array<DailyAggregateCallStatsResponse<CallParticipantCountReport>>}
-   * @memberof CallParticipantCountReportResponse
-   */
-  daily: Array<DailyAggregateCallStatsResponse<CallParticipantCountReport>>;
-}
-
-/**
- *
- * @export
- * @interface CallParticipantCountReport
- */
-export interface CallParticipantCountReport {
-  /**
-   *
-   * @type {Array<ReportByHistogramBucket>}
-   * @memberof CallParticipantCountReport
-   */
-  histogram: Array<ReportByHistogramBucket>;
-}
-
-/**
- *
- * @export
- * @interface CallsPerDayReportResponse
- */
-export interface CallsPerDayReportResponse {
-  /**
-   *
-   * @type {Array<DailyAggregateCallStatsResponse<CallsPerDayReport>>}
-   * @memberof CallsPerDayReportResponse
-   */
-  daily: Array<DailyAggregateCallStatsResponse<CallsPerDayReport>>;
-}
-
-/**
- *
- * @export
- * @interface CallsPerDayReport
- */
-export interface CallsPerDayReport {
-  /**
-   *
-   * @type {number}
-   * @memberof CallsPerDayReport
-   */
-  count: number;
-}
-
-/**
- *
- * @export
- * @interface NetworkMetricsReportResponse
- */
-export interface NetworkMetricsReportResponse {
-  /**
-   *
-   * @type {number}
-   * @memberof NetworkMetricsReportResponse
-   */
-  average_connection_time?: number;
-  /**
-   *
-   * @type {number}
-   * @memberof NetworkMetricsReportResponse
-   */
-  average_jitter?: number;
-  /**
-   *
-   * @type {number}
-   * @memberof NetworkMetricsReportResponse
-   */
-  average_latency?: number;
-  /**
-   *
-   * @type {number}
-   * @memberof NetworkMetricsReportResponse
-   */
-  average_time_to_reconnect?: number;
-}
-
-/**
- *
- * @export
- * @interface DailyAggregateCallStatsResponse
- */
-export interface DailyAggregateCallStatsResponse<T = any> {
-  /**
-   *
-   * @type {string}
-   * @memberof DailyAggregateCallStatsResponse<T>
-   */
-  date: string;
-    /**
-   *
-   * @type {T}
-   * @memberof DailyAggregateCallStatsResponse<T>
-   */
-  report: T;
-}
-
-/**
- *
- * @export
- * @interface QualityScoreReportResponse
- */
-export interface QualityScoreReportResponse {
-  /**
-   *
-   * @type {Array<DailyAggregateCallStatsResponse<QualityScoreReport>>}
-   * @memberof QualityScoreReportResponse
-   */
-  daily: Array<DailyAggregateCallStatsResponse<QualityScoreReport>>;
-}
-
-/**
- *
- * @export
- * @interface QualityScoreReport
- */
-export interface QualityScoreReport {
-  /**
-   *
-   * @type {Array<ReportByHistogramBucket>}
-   * @memberof QualityScoreReport
-   */
-  histogram: Array<ReportByHistogramBucket>;
-}
-
-/**
- *
- * @export
- * @interface ReportByHistogramBucket
- */
-export interface ReportByHistogramBucket {
-  /**
-   *
-   * @type {string}
-   * @memberof ReportByHistogramBucket
-   */
-  category: string;
-  /**
-   *
-   * @type {number}
-   * @memberof ReportByHistogramBucket
-   */
-  count: number;
-  /**
-   *
-   * @type {Bound}
-   * @memberof ReportByHistogramBucket
-   */
-  lower_bound?: Bound;
-  /**
-   *
-   * @type {number}
-   * @memberof ReportByHistogramBucket
-   */
-  mean: number;
-  /**
-   *
-   * @type {number}
-   * @memberof ReportByHistogramBucket
-   */
-  sum: number;
-  /**
-   *
-   * @type {Bound}
-   * @memberof ReportByHistogramBucket
-   */
-  upper_bound?: Bound;
-}
-
-/**
- *
- * @export
- * @interface Bound
- */
-export interface Bound {
-  /**
-   *
-   * @type {boolean}
-   * @memberof Bound
-   */
-  inclusive: boolean;
-  /**
-   *
-   * @type {number}
-   * @memberof Bound
-   */
-  value: number;
-}
-
-/**
- *
- * @export
- * @interface SDKUsageReportResponse
- */
-export interface SDKUsageReportResponse {
-  /**
-   *
-   * @type {Array<DailyAggregateCallStatsResponse<SDKUsageReport>>}
-   * @memberof SDKUsageReportResponse
-   */
-  daily: Array<DailyAggregateCallStatsResponse<SDKUsageReport>>;
-}
-
-/**
- *
- * @export
- * @interface SDKUsageReport
- */
-export interface SDKUsageReport {
-  /**
-   *
-   * @type {{ [key: string]: PerSDKUsageReport; }}
-   * @memberof SDKUsageReport
-   */
-  per_sdk_usage: { [key: string]: PerSDKUsageReport };
-}
-
-/**
- *
- * @export
- * @interface PerSDKUsageReport
- */
-export interface PerSDKUsageReport {
-  /**
-   *
-   * @type {{ [key: string]: number; }}
-   * @memberof PerSDKUsageReport
-   */
-  by_version: { [key: string]: number };
-  /**
-   *
-   * @type {number}
-   * @memberof PerSDKUsageReport
-   */
-  total: number;
-}
-
-/**
- *
- * @export
- * @interface UserFeedbackReportResponse
- */
-export interface UserFeedbackReportResponse {
-  /**
-   *
-   * @type {Array<DailyAggregateCallStatsResponse<UserFeedbackReport>>}
-   * @memberof UserFeedbackReportResponse
-   */
-  daily: Array<DailyAggregateCallStatsResponse<UserFeedbackReport>>;
-}
-
-/**
- *
- * @export
- * @interface UserFeedbackReport
- */
-export interface UserFeedbackReport {
-  /**
-   *
-   * @type {{ [key: string]: number; }}
-   * @memberof UserFeedbackReport
-   */
-  count_by_rating: { [key: string]: number };
-  /**
-   *
-   * @type {number}
-   * @memberof UserFeedbackReport
-   */
-  unreported_count: number;
-}
-
 /**
  * Basic response information
  * @export
@@ -5217,7 +5063,6 @@ export interface QueryAggregateCallStatsResponse {
    */
   user_feedback_report?: UserFeedbackReportResponse;
 }
-
 /**
  *
  * @export
@@ -5606,6 +5451,49 @@ export interface RejectCallResponse {
 /**
  *
  * @export
+ * @interface ReportByHistogramBucket
+ */
+export interface ReportByHistogramBucket {
+  /**
+   *
+   * @type {string}
+   * @memberof ReportByHistogramBucket
+   */
+  category: string;
+  /**
+   *
+   * @type {number}
+   * @memberof ReportByHistogramBucket
+   */
+  count: number;
+  /**
+   *
+   * @type {Bound}
+   * @memberof ReportByHistogramBucket
+   */
+  lower_bound?: Bound;
+  /**
+   *
+   * @type {number}
+   * @memberof ReportByHistogramBucket
+   */
+  mean: number;
+  /**
+   *
+   * @type {number}
+   * @memberof ReportByHistogramBucket
+   */
+  sum: number;
+  /**
+   *
+   * @type {Bound}
+   * @memberof ReportByHistogramBucket
+   */
+  upper_bound?: Bound;
+}
+/**
+ *
+ * @export
  * @interface RequestPermissionRequest
  */
 export interface RequestPermissionRequest {
@@ -5797,6 +5685,32 @@ export interface ScreensharingSettingsResponse {
    * @memberof ScreensharingSettingsResponse
    */
   target_resolution?: TargetResolution;
+}
+/**
+ *
+ * @export
+ * @interface SDKUsageReportResponse
+ */
+export interface SDKUsageReportResponse {
+  /**
+   *
+   * @type {Array<DailyAggregateCallStatsResponse<SDKUsageReport>>}
+   * @memberof SDKUsageReportResponse
+   */
+  daily: Array<DailyAggregateCallStatsResponse<SDKUsageReport>>;
+}
+/**
+ *
+ * @export
+ * @interface SDKUsageReport
+ */
+export interface SDKUsageReport {
+  /**
+   *
+   * @type {{ [key: string]: PerSDKUsageReport; }}
+   * @memberof SDKUsageReport
+   */
+  per_sdk_usage: { [key: string]: PerSDKUsageReport };
 }
 /**
  *
@@ -6666,6 +6580,38 @@ export interface UserDeletedEvent {
    * @memberof UserDeletedEvent
    */
   user?: UserObject;
+}
+/**
+ *
+ * @export
+ * @interface UserFeedbackReportResponse
+ */
+export interface UserFeedbackReportResponse {
+  /**
+   *
+   * @type {Array<DailyAggregateCallStatsResponse<UserFeedbackReport>>}
+   * @memberof UserFeedbackReportResponse
+   */
+  daily: Array<DailyAggregateCallStatsResponse<UserFeedbackReport>>;
+}
+/**
+ *
+ * @export
+ * @interface UserFeedbackReport
+ */
+export interface UserFeedbackReport {
+  /**
+   *
+   * @type {{ [key: string]: number; }}
+   * @memberof UserFeedbackReport
+   */
+  count_by_rating: { [key: string]: number };
+  /**
+   *
+   * @type {number}
+   * @memberof UserFeedbackReport
+   */
+  unreported_count: number;
 }
 /**
  *
